Enable easy running outside of installation directory.
Currently the default checkpoint path specifications assumes that
things are being run from within the installation directory.
This make it hard to run the script from a different working directory.

These edits should keep the current behavior if the checkpoint
specification is indeed a local/relative directory, but will fallback
to the central install (script location) directory if the file is not
present, allowing for running outside of the installation directory.